:recycle: [#4606] Implement variable-case property mappings based on Formik
The layout is still the same, except this table will no longer live in its own
modal, but be a tab in the main modal for the registration options instead.

The fields now make use of Formik's FieldArray to manage a list of nested
objects.
